requirements.txt
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
linkify-it-py==2.0.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psutil==7.0.0
pycparser==2.22
Pygments==2.19.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
textual==2.1.2
textual-fspicker==0.4.1
textual-slider==0.2.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.3.0
vbox-api-soap==2.1.1
websockify==0.12.0
Werkzeug==3.1.3
zeep==4.3.1
